backend/requirements.txt
fastapi==0.118.0
uvicorn==0.37.0
playwright==1.55.0
supabase==2.20.0
websockets==15.0.1
pydantic==2.11.9
python-dotenv==1.1.1
httpx==0.25.0
aiofiles==23.2.1